+++ example/lib/app/pages/showcase/showcase_injections.dart
@@ -1,5 +1,6 @@
 import 'package:get_it/get_it.dart';
 import 'package:nano_core/nano_core.dart';
+import '../../mocks/mock_models.dart';
 import 'showcase_controller.dart';
 
 /// Dependency injection bindings for the showcase page scope.
@@ -9,6 +10,14 @@ class ShowcaseInjections extends NanoInjections {
 
   @override
   void binds(GetIt i) {
-    i.registerFactory<ShowcaseController>(() => ShowcaseController());
+    // NanoHttpClient is resolved automatically from the global AppInjections scope!
+    i.registerLazySingleton<MockUserRepository>(
+      () => MockUserRepository(i<NanoHttpClient>()),
+    );
+    i.registerFactory<ShowcaseController>(
+      () => ShowcaseController(userRepository: i<MockUserRepository>()),
+    );
   }
 }
+
+